HexWorld

HexWorld.cell_materials

HexWorld.cell_properties

HexWorld.Cell

HexWorld.CellData

HexWorld.CellData

HexWorld.CellMaterial

HexWorld.Cubic

HexWorld.EdgeCells

HexWorld.Grid

HexWorld.GridDims

HexWorld.Hex

HexWorld.WallMaterial

Base.iterate

HexWorld.cell_id

HexWorld.cemter_and_vertices_loc

HexWorld.center_loc

HexWorld.create_oval

HexWorld.pack_cell_data

HexWorld.parse_cell_id

HexWorld.cell_materials — Constant

CellWallType

Different wall types

Allowed Values: 1:16

HexWorld.cell_properties — Constant

CellProperty

Generic cell wall properties

HexWorld.Cell — Type

Cell

q: column index, starting from 1

r: row index, starting from 1

Note:

(1,1) represents the upper left corner.

In Hexagons notation this structure would have been called HexagonOffsetOddQ

Units:

length of a hexagon side is 1

width is 2

height is sqrt(3)

HexWorld.CellData — Type

Describes a Hexagonal Cell in a grid.

type::UInt8: points to unique CellMaterial

heigth::UInt8: physical height of the cell

walls::SVector{6,UInt8}: wall meterial IDs from WallMaterial

HexWorld.CellData — Method

CellData(pack)

Unpacks the bytest into CellData structure

HexWorld.CellMaterial — Type

Materials used to describ cell's geometyr and physics.

Each material has an unique id::UInt8 and type::String to identify the type of the material.

HexWorld.Cubic — Type

Cubic

Hexagon represented as the intersection of the unit qube with the plane x + y + z = 1

See Cubic coordinates

HexWorld.EdgeCells — Type

Defines an rectangel inside the gfrid.

HexWorld.Grid — Type

Grid

Rpresents a hexagonal grid. The coordinates are releted to cells as odd-q, see Hexagonal Grids

Params:

dims::GridDims: dimensions of the grid

cells::Array{UInt64,2}: packed cell data

HexWorld.GridDims — Type

Grid Dimensions

Fields:

width: number of columns (q cooridinate)

height: number of rows (r coordinate)

HexWorld.Hex — Type

Hex

Hexagonal cell of a hex grid. Its implementations use different coordinate systesms

See Hexagonal Grids

HexWorld.WallMaterial — Type

WallMaterial describes wall of a grid cell.

id: composed of CellWallType and CellProperty

color: refers to a collor in Color package.

Base.iterate — Function

i terate(dims, state)

Iterates over the grid edge in clockwise direction starting at upper left corener.

HexWorld.cell_id — Method

cell_id(t, p:)

Creates wall id by or-ing type and property bits.

HexWorld.cemter_and_vertices_loc — Method

cemterandvertices_loc(cell)

Returns an 7 by 2 Array{Float64,2} where the first row is the center of the cell and other 6 rows contain vertices.

Vertices are orthered counter-clockwise, starting with the right-most vertex.

HexWorld.center_loc — Method

center_loc(cell)

Returns the (x,y) location of the center of the cell.

HexWorld.create_oval — Method

createoval(f, rect, center:, grid, edgep, inner_p)

Uses rays between each cell on the boundary of the rectangle rect and the center, that is assumed to be inside the rect, and the function f to create a shape in the interior of the rect.

HexWorld.pack_cell_data — Method

packcelldata(c)

Packs the values of a CellData structure into 64bits.

HexWorld.parse_cell_id — Method

parsecellid(w)

Returns Material Type and Property